+++ config/urls.py
@@ -5,7 +5,7 @@ from django.contrib import admin
 from django.views.generic import TemplateView
 from django.views import defaults as default_views
 from statusite.repository.views import repo_list
-
+from statusite.build_events.views import mrbelvedereci_build_webhook
 urlpatterns = [
     # Django Admin, use {% url 'admin:index' %}
     url(settings.ADMIN_URL, admin.site.urls),
@@ -20,6 +20,10 @@ urlpatterns = [
     # repository app
     url(r'^$', repo_list, name='home'),
     url(r'^repo/', include('statusite.repository.urls', namespace='repository')),
+
+    # buildresults api
+    url(r'^webhook/mbci/build$', mrbelvedereci_build_webhook, name="mrbelvedereci_build_webhook"),
+
 
     # django-rq
     url(r'^django-rq/', include('django_rq.urls')),
