+++ statusite/build_events/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class BuildEventsConfig(AppConfig):
+    name = 'build_events'
