+++ statusite/repository/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 from statusite.repository.models import Release
 from statusite.repository.models import Repository
-from statusite.repository.models import BuildResult
 
 class ReleaseAdmin(admin.ModelAdmin):
     list_display = ('repo', 'name', 'beta')
@@ -11,7 +10,3 @@ admin.site.register(Release, ReleaseAdmin)
 class RepositoryAdmin(admin.ModelAdmin):
     list_display = ('name', 'owner')
 admin.site.register(Repository, RepositoryAdmin)
-
-class BuildResultAdmin(admin.ModelAdmin):
-    list_display = ('repo','release','plan_name','status')
-admin.site.register(BuildResult, BuildResultAdmin)
