+++ statusite/repository/models.py
@@ -3,9 +3,10 @@ from __future__ import unicode_literals
 from django.db import models
 from django.urls import reverse
 from django.utils import timezone
+from django.conf import settings
 
 from github3 import login
-from django.conf import settings
+from jsonfield import JSONField
 
 from statusite.repository.exceptions import RepoReloadError
 from statusite.repository.utils import parse_times
@@ -109,20 +110,3 @@ class Release(models.Model):
         return '{}: {}'.format(self.repo.product_name, self.version)
 
 
-class BuildResult(models.Model):
-    class Meta:
-        verbose_name = 'Build Result'
-
-    repo = models.ForeignKey(Repository, related_name='build_results')
-    release = models.ForeignKey(Release, related_name='build_results')
-    plan_name = models.CharField(max_length=255)
-    mbci_build_id = models.IntegerField('mrbelvedereci build id')
-    status = models.CharField(max_length=25)
-    build_date = models.DateField()
-    tests_passed = models.IntegerField()
-    tests_failed = models.IntegerField()
-    tests_total = models.IntegerField()
-    # TODO: additional build stats
-
-    def __str__(self):
-        return '{}: {} [Build {}]'.format(self.repo.product_name, self.release.version, self.mbci_build_id)
